Renamed controller to controllers for consistency
